Shared helper for average score calculation
