+++ package.json
@@ -16,7 +16,8 @@
     "chart.js": "^2.7.2",
     "jquery": "^3.3.1",
     "jquery-serializejson": "^2.8.1",
-    "moment": "^2.22.2"
+    "moment": "^2.22.2",
+    "qs": "^6.5.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.2",
