+++ src/index.html
@@ -66,12 +66,15 @@
 
     $.getJSON("./data/questions.json", function(data) {
       function updateChart() {
-        var averageResults = questions.getAverageGroupResults();
+        var formData = questions.serializeJSON();
+        var averageResults = questions.getAverageGroupResults(formData);
         results.update(averageResults);
       }
 
       var questions = new questionnaire.Questions("#questions", data.questions);
-      var averageResults = questions.getAverageGroupResults();
+      var formData = questions.serializeJSON();
+      console.log(formData);
+      var averageResults = questions.getAverageGroupResults(formData);
       var results = new questionnaire.Results("#results", averageResults);
 
       questions.element.addEventListener("change", updateChart);
