+++ src/questions.html
@@ -17,8 +17,7 @@
     }
 
     h1,
-    h2,
-    h3 {
+    h2 {
       margin-top: 0;
     }
