+++ src/results.html
@@ -17,8 +17,7 @@
     }
 
     h1,
-    h2,
-    h3 {
+    h2 {
       margin-top: 0;
     }
 
@@ -36,26 +35,7 @@
     <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
     <script>
 
-    function getData() {
-      var params = [];
-      var query = window.location.search.substring(1);
-      var vars = query.split("&");
-
-      for (var i = 0; i < vars.length; i++) {
-        var pair = vars[i].split("=");
-        pair[0] = pair[0].replace(/\+/g, "%20");
-
-        params[i] = {
-          name: decodeURIComponent(pair[0]),
-          value: pair[1]
-        };
-      }
-
-      return params;
-    };
-
-    var data = getData();
-    var results = new questionnaire.Results("#results", data);
+    var results = new questionnaire.Results("#results", location.search);
 
     </script>
   </body>
